+++ php/operaciones/consejodirectivo/ci_cd_disposiciones.php
@@ -27,7 +27,8 @@ class ci_cd_disposiciones extends docu_ci
                 // el 27 es para que corte la cadena despues del caracter 27, de /home/fce/consejodirectivo/
                 $nombre = substr($dat['archivo_path'],27);
                 $dir_tmp = toba::proyecto()->get_www_temp();
-                exec("cp '". $dat['archivo_path']. "' '" .$dir_tmp['path']."/".$nombre."'");
+                if(!file_exists($dir_tmp['path']."/".$nombre))
+                    exec("cp '". $dat['archivo_path']. "' '" .$dir_tmp['path']."/".$nombre."'");
                 $temp_archivo = toba::proyecto()->get_www_temp($nombre);
                 $dat['disposicion'] = "<a href='{$temp_archivo['url']}'target='_blank'>".$dat['disposicion_numero']."</a>";    
             }
@@ -57,7 +58,8 @@ class ci_cd_disposiciones extends docu_ci
             // el 19 es para que corte la cadena despues del caracter 19, de /home/fce/consejodirectivo/
             $nombre = substr($datos['archivo_path'],27);
             $dir_tmp = toba::proyecto()->get_www_temp();
-            exec("cp '". $datos['archivo_path']. "' '" .$dir_tmp['path']."/".$nombre."'");
+            if(!file_exists($dir_tmp['path']."/".$nombre))
+                exec("cp '". $datos['archivo_path']. "' '" .$dir_tmp['path']."/".$nombre."'");
             $temp_archivo = toba::proyecto()->get_www_temp($nombre);
             $tamanio = round(filesize($temp_archivo['path']) / 1024);
             $datos['archivo_path'] = "<a href='{$temp_archivo['url']}'target='_blank'>Descargar archivo</a>";
